backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
pytest==8.3.2
pytest-django==4.8.0
sqlparse==0.5.1
